# Conflicts:
#	pom.xml
#	src/main/java/com/baibaoxiang/controller/ArticleController.java
#	src/main/java/com/baibaoxiang/mapper/AreaMapper.java
#	src/main/java/com/baibaoxiang/mapper/AreaMapper.xml
#	src/main/java/com/baibaoxiang/mapper/ArticleMapper.xml
#	src/main/java/com/baibaoxiang/mapper/ManagerMapper.xml
#	src/main/java/com/baibaoxiang/mapper/SchoolMapper.xml
#	src/main/java/com/baibaoxiang/mapper/custom/ArticleMapperCustom.xml
#	src/main/java/com/baibaoxiang/po/Area.java
#	src/main/java/com/baibaoxiang/po/Article.java
#	src/main/java/com/baibaoxiang/po/ArticleType.java
#	src/main/java/com/baibaoxiang/po/Manager.java
#	src/main/java/com/baibaoxiang/po/School.java
#	src/main/java/com/baibaoxiang/service/AreaService.java
#	src/main/java/com/baibaoxiang/service/ArticleService.java
#	src/main/java/com/baibaoxiang/service/SchoolService.java
#	src/main/java/com/baibaoxiang/serviceimpl/ArticleServiceImpl.java
#	src/main/java/com/baibaoxiang/serviceimpl/SchoolServiceImpl.java
#	src/main/java/com/baibaoxiang/serviceimpl/SearchServiceImpl.java
#	src/main/java/com/baibaoxiang/tool/FastDfsClient.java
#	src/main/resources/spring/applicationContext-service.xml

+++ src/main/java/com/baibaoxiang/mapper/AreaMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface AreaMapper {
-    int countByExample(AreaExample example);
+    long countByExample(AreaExample example);
 
     int deleteByExample(AreaExample example);
